Add configuration for the nexus-staging-maven-plugin.
This plugin requires to be declared with extension=true.
It is relatively intrusive, thus isolating it inside the oss-release profile.

The plugin also has "rc" goals that can be used with mvn cli invokation
 to automate the nexus workflow. Using the oss-release for this usecase
 is inconvenient, providing a separate profile "nexus-plugin" to enable
 the plugin without the side-effects of the oss-release profile.

Configure the maven-deploy-plugin to be skipped so that it does not
 interfere with the nexus-staging-maven-plugin. Also configure the
 nexus-staging-maven-plugin to skip its execution based on maven.deploy.skip
 which is a well known property.

Add comments to clarify what each profile does.